putting signatures directly into code, even though it's ugly as sin
